alter model fields for extra functionality
* `object_id` field altered from `PositiveInteger` to `TextField`

`content_object` `GenericForeignKey` may have
either `string` or `number` values. So now `object_id`
which is generated from `content_type` is able
to accept both types of values saved as a `string`.

* removed `content_type`s `NOT NULL` constraint

Sometimes there is a need to log an action
and describe the event by storing values
to extra_data but doesn't involve an object.

* `content_type` replaced on_delete=CASCADE to on_delete=SET_NULL

log entries that relates to an object must be kept to
the database when the related object is deleted.

* added tranlatable field names